requirements.txt
django >=3.2.4, <4.0
django-bootstrap4
django-cleanup
django-environ
django-fontawesome-5
django-recaptcha2
factory-boy
gunicorn
pillow
psycopg2
pytest
pytest-django
